+++ apps/cms/src/payload-types.ts
@@ -117,7 +117,6 @@ export interface Config {
   };
   locale: null;
   widgets: {
-    'season-sync': SeasonSyncWidget;
     collections: CollectionsWidget;
   };
   user: User;
@@ -1021,16 +1020,6 @@ export interface SeasonSelect<T extends boolean = true> {
   updatedAt?: T;
   createdAt?: T;
   globalType?: T;
-}
-/**
- * This interface was referenced by `Config`'s JSON-Schema
- * via the `definition` "season-sync_widget".
- */
-export interface SeasonSyncWidget {
-  data?: {
-    [k: string]: unknown;
-  };
-  width: 'medium' | 'large' | 'x-large' | 'full';
 }
 /**
  * This interface was referenced by `Config`'s JSON-Schema